Enables parsing a page when site_info has no namespaces
The default header/footer used by Dump.from_page_xml to wrap a <page> element in
a bare-bones <mediawiki> element doesn't contain any namespaces. That means that
the namespace_map in Dump.from_element is None.

When you start iterating through Dump.pages, that eventually leads to a
TypeError in page.extract_namespace.

This commit just makes {} the fallback value for the namespace map to fix that
particular issue.